+++ core/src/test/java/io/grpc/util/ForwardingLoadBalancerHelperTest.java
@@ -14,10 +14,13 @@
  * limitations under the License.
  */
 
-package io.grpc;
+package io.grpc.util;
 
 import static org.mockito.Mockito.mock;
 
+import io.grpc.EquivalentAddressGroup;
+import io.grpc.ForwardingTestUtil;
+import io.grpc.LoadBalancer;
 import java.lang.reflect.Method;
 import java.net.SocketAddress;
 import java.util.Arrays;
